soapy: Use the new IOCTL call on enumeration soapy: Si5351: Move device ID/revision support checks to constructor instead of init function soapy: Si5351: Normalize phase to 0-360 instead of throwing an error soapy: Si5351: Add method to mask/unmask a clock enable from the OEb pin soapy: Si5351: Add IRQ config soapy: Set external clock output independent of OEb pin soapy: Add AuxMCU base class soapy: Add PMC MCU communication library barebones soapy: Add LT7182S support library soapy: Add 8V97003 mmWave synth support library soapy: Add preliminary ExpansionCard base class soapy: AXIIIC: count is uint8_t, limited by the controller dynamic mode, for now soapy: Misc fixes

+++ software/soapy/src/include/AXIIIC.hpp
@@ -103,11 +103,11 @@ public:
 
     void transmit(uint8_t address, uint8_t *buf, uint8_t count, AXIIIC::Stop stop = AXIIIC::Stop::STOP);
 
-    inline void write(uint8_t address, uint8_t *src, uint32_t count, AXIIIC::Stop stop = AXIIIC::Stop::STOP)
+    inline void write(uint8_t address, uint8_t *src, uint8_t count, AXIIIC::Stop stop = AXIIIC::Stop::STOP)
     {
         this->transmit((address << 1) & ~0x01, src, count, stop);
     }
-    inline void read(uint8_t address, uint8_t *dst, uint32_t count, AXIIIC::Stop stop = AXIIIC::Stop::STOP)
+    inline void read(uint8_t address, uint8_t *dst, uint8_t count, AXIIIC::Stop stop = AXIIIC::Stop::STOP)
     {
         this->transmit((address << 1) | 0x01, dst, count, stop);
     }
